azure_utilities/azure_functions_python/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-mgmt-resource
azure-common
email-to
msrestazure